+++ Code/Commando/ServerSettings.h
@@ -59,7 +59,7 @@ class ServerSettingsClass
 		static void Encrypt_Serial(StringClass serial_in, StringClass &serial_out, bool encrypt = true);
 		static void Decrypt_Serial(StringClass serial_in, StringClass &serial_out);
 		static const char *Get_Master_Server_Password(void) {return(MasterPassword);}
-		static unsigned long Get_Master_Bandwidth(void) {return(MasterBandwidth);}
+		static unsigned int Get_Master_Bandwidth(void) {return(MasterBandwidth);}
 		static bool Is_Active(void) {return(IsActive);}
 		static char *Get_Settings_File_Name(void) {return(SettingsFile);}
 		static bool Check_Game_Settings_File(char *config_file);
@@ -76,7 +76,7 @@ class ServerSettingsClass
 		static bool IsActive;
 
 		static char MasterPassword[128];
-		static unsigned long MasterBandwidth;
+		static unsigned int MasterBandwidth;
 		static char PreferredLoginServer[256];
 		static GameModeTypeEnum GameMode;
 		static int DiskLogSize;
