+++ Code/Commando/dlgmpslaveservers.cpp
@@ -47,7 +47,7 @@
 
 
 
-unsigned long SlaveServerDialogClass::EnableIDs[MAX_SLAVES] = {
+unsigned int SlaveServerDialogClass::EnableIDs[MAX_SLAVES] = {
 	IDC_SLAVE_ENABLE1,
 	IDC_SLAVE_ENABLE2,
 	IDC_SLAVE_ENABLE3,
@@ -58,7 +58,7 @@ unsigned long SlaveServerDialogClass::EnableIDs[MAX_SLAVES] = {
 };
 
 
-unsigned long SlaveServerDialogClass::NickIDs[MAX_SLAVES] = {
+unsigned int SlaveServerDialogClass::NickIDs[MAX_SLAVES] = {
 	IDC_NICK_EDIT1,
 	IDC_NICK_EDIT2,
 	IDC_NICK_EDIT3,
@@ -68,7 +68,7 @@ unsigned long SlaveServerDialogClass::NickIDs[MAX_SLAVES] = {
 	IDC_NICK_EDIT7
 };
 
-unsigned long SlaveServerDialogClass::PassIDs[MAX_SLAVES] = {
+unsigned int SlaveServerDialogClass::PassIDs[MAX_SLAVES] = {
 	IDC_PASS_EDIT1,
 	IDC_PASS_EDIT2,
 	IDC_PASS_EDIT3,
@@ -79,7 +79,7 @@ unsigned long SlaveServerDialogClass::PassIDs[MAX_SLAVES] = {
 };
 
 
-unsigned long SlaveServerDialogClass::SerialIDs[MAX_SLAVES] = {
+unsigned int SlaveServerDialogClass::SerialIDs[MAX_SLAVES] = {
 	IDC_SERIAL_EDIT1,
 	IDC_SERIAL_EDIT2,
 	IDC_SERIAL_EDIT3,
@@ -91,7 +91,7 @@ unsigned long SlaveServerDialogClass::SerialIDs[MAX_SLAVES] = {
 
 
 
-unsigned long SlaveServerDialogClass::PortIDs[MAX_SLAVES] = {
+unsigned int SlaveServerDialogClass::PortIDs[MAX_SLAVES] = {
 	IDC_PORT_EDIT1,
 	IDC_PORT_EDIT2,
 	IDC_PORT_EDIT3,
@@ -103,7 +103,7 @@ unsigned long SlaveServerDialogClass::PortIDs[MAX_SLAVES] = {
 
 
 
-unsigned long SlaveServerDialogClass::SettingsButtons[MAX_SLAVES] = {
+unsigned int SlaveServerDialogClass::SettingsButtons[MAX_SLAVES] = {
 	IDC_SLAVE1_SETTINGS,
 	IDC_SLAVE2_SETTINGS,
 	IDC_SLAVE3_SETTINGS,
@@ -182,7 +182,7 @@ SlaveServerDialogClass::On_Init_Dialog (void)
 		Set_Dlg_Item_Text(SerialIDs[i], tempser.Peek_Buffer());
 
 		char temp[32];
-		_itoa((unsigned long)port, temp, 10);
+		_itoa((unsigned int)port, temp, 10);
 		WideStringClass tempport(temp, true);
 		Set_Dlg_Item_Text(PortIDs[i], tempport.Peek_Buffer());
 
@@ -195,7 +195,7 @@ SlaveServerDialogClass::On_Init_Dialog (void)
 
 
 void
-SlaveServerDialogClass::On_Command (int ctrl_id, int message_id, DWORD param)
+SlaveServerDialogClass::On_Command (int ctrl_id, int message_id, unsigned int param)
 {
 	switch (ctrl_id) {
 		case IDC_MENU_BACK_BUTTON:
@@ -227,7 +227,7 @@ SlaveServerDialogClass::On_Command (int ctrl_id, int message_id, DWORD param)
 				WideStringClass tempport(Get_Dlg_Item_Text(PortIDs[i]));
 				StringClass newport;
 				tempport.Convert_To(newport);
-				unsigned long port = atoi(newport.Peek_Buffer());
+				unsigned int port = atoi(newport.Peek_Buffer());
 				if (port > 0xffff) {
 					port = 0xffff;
 				}
